requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0  # Optional: for environment variable management
requests>=2.31.0  # Already included with openai, but listed explicitly